fix(ui): 将易读/保守排版开关独立为「布局」区块并置顶
原开关藏在「外观」卡片深处不易发现；现于设置→通用页录音输入下方展示独立「布局」卡片。
