+++ openless-all/app/src/i18n/ja.ts
@@ -1299,6 +1299,9 @@ export const ja: typeof zhCN = {
       ko: '한국어 (Beta)',
       restartHint: '一部のネイティブメニュー（トレイ等）は再起動後に反映されます。',
     },
+    layout: {
+      title: 'レイアウト',
+    },
     theme: {
       title: '外観',
       label: 'テーマ',
