+++ openless-all/app/src/i18n/ko.ts
@@ -1299,6 +1299,9 @@ export const ko: typeof zhCN = {
       ko: '한국어 (Beta)',
       restartHint: '일부 네이티브 메뉴(트레이 등)는 앱 재시작 후 반영될 수 있습니다.',
     },
+    layout: {
+      title: '레이아웃',
+    },
     theme: {
       title: '모양',
       label: '테마',
